+++ ppr-ui/src/composables/index.ts
@@ -1,2 +1,1 @@
-export * from './useDateHelper'
 export * from './useInputRules'
