+++ ppr-ui/src/utils/date-helper.ts
@@ -1,5 +1,6 @@
 import moment, { Moment } from 'moment'
 import 'moment-timezone'
+import { isDate } from 'lodash'
 
 /** returns timstamp string in 12 hour format */
 export function format12HourTime (date: Date): string {
@@ -69,4 +70,91 @@ export function isInt (intValue) {
   }
   var x = parseFloat(intValue)
   return (x | 0) === x
+}
+
+/**
+ * Creates and returns a new Date object in UTC, given parameters in Pacific timezone.
+ * (This works regardless of user's local clock/timezone.)
+ * @example "2021, 0, 1, 0, 0" -> "2021-01-01T08:00:00.000Z"
+ * @example "2021, 6, 1, 0, 0" -> "2021-07-01T07:00:00.000Z"
+ */
+export function createUtcDate (year: number, month: number, day: number, hours: number = 0, minutes: number = 0): Date {
+  // FUTURE: change this to get the date from the server
+  const jsDate = new Date()
+  const date = new Date(jsDate.toLocaleString('en-US', { timeZone: 'America/Vancouver' }))
+
+  // update all date and time fields
+  date.setFullYear(year, month, day)
+  date.setHours(hours, minutes, 0, 0) // zero out seconds and milliseconds
+
+  return date
+}
+
+/**
+ * Converts a date string (YYYY-MM-DD) to a Date object at 12:00:00 am Pacific time.
+ * @example 2021-11-22 -> 2021-11-22T08:00:00.00Z
+ */
+export function yyyyMmDdToDate (dateStr: string): Date {
+  // safety checks
+  if (!dateStr) return null
+  if (dateStr.length !== 10) return null
+
+  const split = dateStr.split('-')
+  const year = +split[0]
+  const month = +split[1]
+  const day = +split[2]
+
+  return createUtcDate(year, (month - 1), day)
+}
+
+/**
+ * Converts a Date object to a date string (YYYY-MM-DD) in Pacific timezone.
+ * @example "2021-01-01 07:00:00 GMT" -> "2020-12-31"
+ * @example "2021-01-01 08:00:00 GMT" -> "2021-01-01"
+ */
+export function dateToYyyyMmDd (date: Date): string {
+  // safety check
+  if (!isDate(date) || isNaN(date.getTime())) return null
+
+  return date.toLocaleDateString('en-CA', {
+    timeZone: 'America/Vancouver',
+    month: 'numeric', // 12
+    day: 'numeric', // 31
+    year: 'numeric' // 2020
+  })
+}
+
+/**
+ * Converts a date string (YYYY-MM-DD) to a date string (Month Day, Year) in Pacific timezone.
+ * @param longMonth whether to show long month name (eg, December vs Dec)
+ * @param showWeekday whether to show the weekday name (eg, Thursday)
+ * @example "2021-01-01" -> "Thursday, December 31, 2020"
+ */
+export function yyyyMmDdToPacificDate (dateStr: string, longMonth = false, showWeekday = false): string {
+  return dateToPacificDate(yyyyMmDdToDate(dateStr), longMonth, showWeekday)
+}
+
+/**
+ * Converts a Date object to a date string (Month Day, Year) in Pacific timezone.
+ * @param longMonth whether to show long month name (eg, December vs Dec)
+ * @param showWeekday whether to show the weekday name (eg, Thursday)
+ * @example "2021-01-01 07:00:00 GMT" -> "Dec 31, 2020"
+ * @example "2021-01-01 08:00:00 GMT" -> "Jan 1, 2021"
+ */
+export function dateToPacificDate (date: Date, longMonth = false, showWeekday = false): string {
+  // safety check
+  if (!isDate(date) || isNaN(date.getTime())) return null
+
+  let dateStr = date.toLocaleDateString('en-CA', {
+    timeZone: 'America/Vancouver',
+    weekday: showWeekday ? 'long' : undefined, // Thursday or nothing
+    month: longMonth ? 'long' : 'short', // December or Dec
+    day: 'numeric', // 31
+    year: 'numeric' // 2020
+  })
+
+  // remove period after month
+  dateStr = dateStr.replace('.', '')
+
+  return dateStr
 }
